basic level setup
made few changes in the scenes and added few new scene along with that created the images in the scene.
next step is to add buttons to the images. there is also new script made for the level selection page but its incomplete and the template is ready for the code.