NEW MENU
made a new scene for the menu.
older on is good but functional anymore.
to solve it i will do the changes in this sample scene and once it works properly all the changes can be shifted to the other scene or will make this scene as the og menu page.